Moved MarkupsToModel extension from SlicerIGT
This module was originally in https://github.com/SlicerIGT/SlicerIGT but it proved to be useful for many other tasks,
not just for IGT applications. Moved to a separate extension to allow other modules that need model from markup
to use this module without depending on the rather large SlicerIGT extension.